+++ core-bundle/src/Resources/contao/pages/PageRegular.php
@@ -400,7 +400,7 @@ class PageRegular extends \Frontend
 			{
 				try
 				{
-					$version = PackageUtil::getVersion('contao-components/jquery');
+					$version = PackageUtil::getNormalizedVersion('contao-components/jquery');
 
 					$this->Template->mooScripts .= \Template::generateScriptTag('https://code.jquery.com/jquery-' . $version . '.min.js') . "\n";
 
@@ -428,7 +428,7 @@ class PageRegular extends \Frontend
 			{
 				try
 				{
-					$version = PackageUtil::getVersion('contao-components/jquery');
+					$version = PackageUtil::getNormalizedVersion('contao-components/mootools');
 
 					if (version_compare($version, '1.5.1', '>'))
 					{